Added JSONLoader class and dependencies.
Removed jeayeson. That was stupid. Using jsoncpp lib instead.
JSONLoader should be used to read and write JSON data.
The data passed to JSONLoader should have a static memory
location. The data will be written back to JSON from the
same location.

The warnings is stupid and I don't know how to fix it yet.